requirements.txt
amqp==1.4.9
anyjson==0.3.3
Babel==2.2.0
backports.ssl-match-hostname==3.5.0.1
billiard==3.3.0.22
celery==3.1.20
certifi==2015.11.20.1
Django==1.9.1
django-celery==3.1.17
django-filter==0.12.0
django-push-notifications==1.4.1
djangorestframework==3.3.2
flower==0.8.4
futures==3.0.5
gunicorn==19.4.5
kombu==3.0.33
Markdown==2.6.5
MySQL-python==1.2.5
Pillow==2.9.0
pytz==2015.7
tornado==4.2
wheel==0.24.0

